requirements.txt
# Plant Disease Classification - Requirements
# GPU: NVIDIA GeForce RTX 3050 Laptop
# CUDA Version: 13.0 (using PyTorch CUDA 12.4 - backward compatible)

# ============================================
# INSTALLATION INSTRUCTIONS
# ============================================
# Step 1: Uninstall existing PyTorch (if any)
#   pip uninstall torch torchvision torchaudio
#
# Step 2: Install PyTorch with CUDA 12.4 support
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu124
#
# Step 3: Install remaining dependencies
#   pip install -r requirements.txt
#
# Step 4: Verify CUDA is working
#   python -c "import torch; print(f'CUDA: {torch.cuda.is_available()}')"
#
# ============================================

# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.65.0
gdown
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision

# Web Application
streamlit>=1.28.0

# Logging and Visualization
tensorboard>=2.13.0